[MIG]cetmix_tower_server: migrate ace_tower widget from 16.0
Task: 4865
