+++ cetmix_tower_server/__manifest__.py
@@ -76,7 +76,7 @@
     ],
     "assets": {
         "web.assets_backend": [
-            "cetmix_tower_server/static/src/**/*.xml",
+            "cetmix_tower_server/static/src/components/**/*.xml",
             "cetmix_tower_server/static/src/**/*.js",
             "cetmix_tower_server/static/src/**/*.scss",
         ],
