+++ cetmix_tower_server/static/src/components/ace_variables/ace_variables.scss
@@ -13,8 +13,8 @@
         min-height: 200px !important;
         height: auto !important;
         width: 100% !important;
-        border: 1px solid #ced4da;
-        border-radius: 4px;
+        border: 1px solid var(--bs-border-color);
+        border-radius: var(--bs-border-radius);
         display: block !important;
     }
