+++ cetmix_tower_server/static/src/components/ace_variables/autocomplete_popup.xml
@@ -45,6 +45,7 @@
                 t-ref="itemsContainer"
                 t-att-id="props.type === 'secrets' ? 'secrets-list' : 'variables-list'"
                 role="listbox"
+                aria-live="polite"
             >
                 <div
                     t-foreach="filteredCommands"
