+++ cetmix_tower_server/views/cx_tower_variable_value_view.xml
@@ -21,7 +21,11 @@
                     name="plan_line_action_id"
                     readonly="server_id or server_template_id"
                 />
-                <field name="value_char" readonly="variable_type == 'o'" widget="ace_tower" />
+                <field
+                    name="value_char"
+                    readonly="variable_type == 'o'"
+                    widget="ace_tower"
+                />
                 <field
                     name="option_id"
                     optional="show"
